Rename transcript to summary since it has only summary endpoints now
