+++ src/web/routes.rs
@@ -4,12 +4,12 @@ use axum::{Router, routing::get};
 use sqlx::PgPool;
 use tower_http::{services::ServeDir, trace::TraceLayer};
 
-pub mod transcript;
+pub mod summary;
 
 pub fn routes(pool: PgPool) -> Router {
 	Router::new()
 		.route("/", get(|| async { "hello world" }))
-		.nest("/summary", transcript::routes())
+		.nest("/summary", summary::routes())
 		// layers run from bottom to top
 		.fallback_service(ServeDir::new(
 			env::var("PUBLIC_DIR").unwrap_or("public/".into()),
