+++ src/web/routes/summary.rs
